CRUD
Added Project CRUD functions

+++ ProjectTracker/Controllers/ProjectsController.cs
@@ -50,7 +50,7 @@ namespace ProjectTracker.Controllers
         // GET: Projects/Create
         public IActionResult Create()
         {
-            ViewData["ClientID"] = new SelectList(_context.Clients, "ClientID", "ClientID");
+            ViewData["ClientID"] = new SelectList(_context.Clients, "ClientID", "ClientName");
             return View();
         }
 
@@ -59,15 +59,25 @@ namespace ProjectTracker.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ProjectID,ClientID,ProjectName,Status,EnteredBy,Notes,LastUpdated,DateEntered")] Project project)
+        public async Task<IActionResult> Create([Bind("ClientID,ProjectName,Status,EnteredBy,Notes,DateEntered")] Project project)
         {
-            if (ModelState.IsValid)
+            try
             {
-                _context.Add(project);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    project.DateEntered = DateTime.Now;
+                    project.LastUpdated = DateTime.Now;
+                    _context.Add(project);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+                ViewData["ClientID"] = new SelectList(_context.Clients, "ClientID", "ClientName", project.ClientID);
             }
-            ViewData["ClientID"] = new SelectList(_context.Clients, "ClientID", "ClientID", project.ClientID);
+            catch (DbUpdateException /* ex */)
+            {
+                //Log the error (uncomment ex variable name and write a log.
+                ModelState.AddModelError("", "Unable to save changes. Try again, and if the problem persists contact your system administrator.");
+            }
             return View(project);
         }
 
@@ -84,48 +94,44 @@ namespace ProjectTracker.Controllers
             {
                 return NotFound();
             }
-            ViewData["ClientID"] = new SelectList(_context.Clients, "ClientID", "ClientID", project.ClientID);
+            ViewData["ClientID"] = new SelectList(_context.Clients, "ClientID", "ClientName", project.ClientID);
             return View(project);
         }
 
         // POST: Projects/Edit/5
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
+        [HttpPost, ActionName("Edit")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ProjectID,ClientID,ProjectName,Status,EnteredBy,Notes,LastUpdated,DateEntered")] Project project)
+        public async Task<IActionResult> EditPost(int? id)
         {
-            if (id != project.ProjectID)
+            if (id == null)
             {
                 return NotFound();
             }
 
-            if (ModelState.IsValid)
+            var projectToUpdate = await _context.Projects.SingleOrDefaultAsync(p => p.ProjectID == id);
+            if (await TryUpdateModelAsync<Project>(projectToUpdate, "", p => p.ClientID, p => p.ProjectName, p => p.Status, p => p.EnteredBy, p => p.Notes, p => p.LastUpdated))
             {
+
                 try
                 {
-                    _context.Update(project);
+                    projectToUpdate.LastUpdated = DateTime.Now;
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (DbUpdateException /* ex */)
                 {
-                    if (!ProjectExists(project.ProjectID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    //Log the error (uncomment ex variable name and write a log.)
+                    ModelState.AddModelError("", "Unable to save changes. Try again, and if the problem persists contact your system administrator.");
                 }
-                return RedirectToAction(nameof(Index));
             }
-            ViewData["ClientID"] = new SelectList(_context.Clients, "ClientID", "ClientID", project.ClientID);
-            return View(project);
+            ViewData["ClientID"] = new SelectList(_context.Clients, "ClientID", "ClientName", projectToUpdate.ClientID);
+            return View(projectToUpdate);
         }
 
         // GET: Projects/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -134,10 +140,18 @@ namespace ProjectTracker.Controllers
 
             var project = await _context.Projects
                 .Include(p => p.Client)
+                .AsNoTracking()
                 .SingleOrDefaultAsync(m => m.ProjectID == id);
             if (project == null)
             {
                 return NotFound();
+            }
+
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                    "Delete failed. Try again, and if the problem persists " +
+                    "see your system administrator.";
             }
 
             return View(project);
@@ -148,10 +162,25 @@ namespace ProjectTracker.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var project = await _context.Projects.SingleOrDefaultAsync(m => m.ProjectID == id);
-            _context.Projects.Remove(project);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            var project = await _context.Projects
+                .AsNoTracking()
+                .SingleOrDefaultAsync(m => m.ProjectID == id);
+            if (project == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
+            try
+            {
+                _context.Projects.Remove(project);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException /* ex */)
+            {
+                //Log the error (uncomment ex variable name and write a log.)
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
         }
 
         private bool ProjectExists(int id)
